refactor(All): profit 자료형 int로 변경

+++ src/main/java/dto/BlackjackProfitDto.java
@@ -3,7 +3,7 @@ package dto;
 import java.util.List;
 
 public record BlackjackProfitDto(
-    double dealerProfit,
+    int dealerProfit,
     List<PlayerResultDto> playerResultDtoList
 ) {
